refactor/type-n-dto
